Update spec to use the CollectionJSON parser
Always parsing the items directly without having a wrapper around
the Colleciton+JSON made the code a bit harder to read with all
the finds that had to be performed. This way it's much easier.

For example, lets compare retrieving the href for an item that has
the title 'Inbox':

    parsed_json['collection']['items'].first['data].find { |data_field|
      data_field['name'] == 'title' && data_field['value'] == 'Inbox'
    }['href']

verses:

    parsed_collection.items.find { |item|
      item.data['title'].value == 'Inbox'
    }.href

which I think looks much better and is easier to read.